+++ packages/rust-bindings/packages/wallet/src/external_functions.rs
@@ -3,6 +3,17 @@ use concordium_rust_bindings_common::{
     helpers::{to_js_error, JsResult},
     types::{Base58String, HexString, JsonString},
 };
+use wallet_library::{
+    credential::create_unsigned_credential_v1_aux,
+    identity::{create_identity_object_request_v1_aux, create_identity_recovery_request_aux},
+    wallet::{
+        get_account_public_key_aux, get_account_signing_key_aux,
+        get_attribute_commitment_randomness_aux, get_credential_id_aux, get_id_cred_sec_aux,
+        get_prf_key_aux, get_signature_blinding_randomness_aux,
+        get_verifiable_credential_backup_encryption_key_aux,
+        get_verifiable_credential_public_key_aux, get_verifiable_credential_signing_key_aux,
+    },
+};
 use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen(js_name = generateUnsignedCredential)]
@@ -42,7 +53,8 @@ pub fn get_credential_deployment_info_ext(signatures: &JsValue, unsigned_info: &
 
 #[wasm_bindgen(js_name = createIdRequestV1)]
 pub fn create_id_request_v1_ext(input: JsonString) -> JsResult {
-    create_id_request_v1_aux(serde_json::from_str(&input).unwrap()).map_err(to_js_error)
+    create_identity_object_request_v1_aux(serde_json::from_str(&input).unwrap())
+        .map_err(to_js_error)
 }
 
 #[wasm_bindgen(js_name = createIdentityRecoveryRequest)]